move computational_form to ssnl
