+++ pysat/utils/_core.py
@@ -36,38 +36,6 @@ def set_data_dir(path=None, store=True):
         raise ValueError('Path %s does not lead to a valid directory.' % path)
 
 
-def computational_form(data):
-    """
-    Repackages numbers, Series, or DataFrames
-
-    Regardless of input format, mathematical operations may be performed on the output via the same pandas mechanisms.
-
-    This method may be particularly useful in analysis methods that aim to be instrument independent. pysat.Instrument objects can package data in a variety of ways within a DataFrame, depending upon the scientific data source. Thus, a variety of data types will be encountered by instrument independent methods and computational_form method may reduce the effort required to support more generalized processing.
-
-    Parameters
-    ----------
-    data : pandas.Series
-        Series of numbers, Series, DataFrames
-
-    Returns
-    -------
-    pandas.Series, DataFrame, or Panel
-        repacked data, aligned by indices, ready for calculation
-    """
-
-    from pysat import DataFrame, Series, datetime, Panel
-
-    if isinstance(data.iloc[0], DataFrame):
-        dslice = Panel.from_dict(dict([(i, data.iloc[i])
-                                       for i in xrange(len(data))]))
-    elif isinstance(data.iloc[0], Series):
-        dslice = DataFrame(data.tolist())
-        dslice.index = data.index
-    else:
-        dslice = data
-    return dslice
-
-
 def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
                  epoch_name='Epoch', units_label='units',
                  name_label='long_name', notes_label='notes',
@@ -304,7 +272,7 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
                         loop_list.append(loop_frame.iloc[step_size*i:step_size*(i+1)])
                         loop_list[-1].index = new_index[step_size*i:step_size*(i+1)]
                         loop_list[-1].index.name = new_index_name
-                
+
                 # add 2D object data, all based on a unique dimension within
                 # netCDF, to loaded data dictionary
                 loadedVars[obj_key_name] = loop_list
